refactor: pr requested changes
