+++ src/types.ts
@@ -1,3 +1,4 @@
 export type Longtitude = number;
 export type Latitude = number;
 export type Zoom = number;
+export type Limits = [number, number, number, number];
