--- tests/configurations/integration/jest.config.js
@@ -1,21 +1,0 @@
-module.exports = {
-  transform: {
-    '^.+\\.ts$': 'ts-jest',
-  },
-  coverageReporters: ['text', 'html'],
-  rootDir: '../../../.',
-  testMatch: ['<rootDir>/tests/integration/**/*.spec.ts'],
-  setupFiles: ['<rootDir>/tests/configurations/jest.setup.js'],
-  collectCoverage: true,
-  moduleDirectories: ['node_modules', 'src'],
-  collectCoverageFrom: ['<rootDir>/src/**/*.{ts}', '!**/node_modules/**', '!**/vendor/**'],
-  preset: 'ts-jest',
-  reporters: [
-    'default',
-    [
-      'jest-html-reporters',
-      { multipleReportsUnitePath: './reports', pageTitle: 'integration', publicPath: './reports', filename: 'integration.html' },
-    ],
-  ],
-  testEnvironment: 'node',
-};
